Changed overlap issue and removed drawings
Was able to drawings and logo overlap. However, I am not sure how to go about the blurry issue. I don't notice it being blurry on my side. I did try to resize the clique logo to see if that would help and for me it made it worse. Do you have any ideas?